refactor: 💡Move some enums to quickpay package
