--- apps/api/src/main/java/dk/treecreate/api/order/Currency.java
@@ -1,13 +1,0 @@
-package dk.treecreate.api.order;
-
-public enum Currency
-{
-    DKK("dkk");
-
-    public final String label;
-
-    Currency(String label)
-    {
-        this.label = label;
-    }
-}
